Refactor command line argument parsing
