+++ reports/annual_min_max_report.py
@@ -1,6 +1,6 @@
-from data_manager import DataManager
-from data_manager.Day import Day
-from data_manager.Year import Year
+from data_manager.data_manager import DataManager
+from data_manager.day import Day
+from data_manager.year import Year
 from reports.report import Report
