+++ reports/report.py
@@ -1,4 +1,4 @@
-from data_manager import DataManager
+from data_manager.data_manager import DataManager
 
 
 class Report:
